+++ Algorithms and data structures/Fibonacci.cs
@@ -8,7 +8,7 @@ namespace Algorithms_and_data_structures
 {
     class Fibonacci
     {
-        public static int fib_recursive(int n)
+        public static ulong fib_recursive(ulong n)
         {
             if ((n == 1) || (n == 2))
                 return 1;
@@ -16,16 +16,16 @@ namespace Algorithms_and_data_structures
                 return fib_recursive(n - 1) + fib_recursive(n - 2);
         }
 
-        public static int fib_iteratively(int m)
+        public static ulong fib_iteratively(ulong m)
         {
             if (m <= 2)
                 return 1;
             else
             {
-                int f1 = 1;
-                int f2 = 1;
-                int temp;
-                for (int i = 3; i <= m; i++)
+                ulong f1 = 1;
+                ulong f2 = 1;
+                ulong temp;
+                for (ulong i = 3; i <= m; i++)
                 {
                     temp = f1 + f2;
                     f1 = f2;
